requirements.txt
twisted>=16.0
python-geoip
python-geoip-geolite2
dataset>=0.7.1
